Code/Udemy_DeepCV/code/24. GANS_Generative_Networks/Face-Aging-with-Identity-Preserved-Conditional-Generative-Adversarial-Networks-master/requirements.txt
tensorflow-gpu==1.4.1
scipy==1.0.0
opencv-python==3.3.0.10
numpy==1.11.0
Pillow==5.1.0
